// apps/buyer-keyword/src/app/shared/interfaces/buyer.ts
export interface Buyer {
  bounceBackMessage?: string;
  buyerKeyword?: string;
  buyerKeywordId?: string;
  buyerName?: string;
  createdDateTime?: string;
  displayPredictedValues?: string;
  forwardingEmailAddress?: string;
  indexingSolutionId?: string;
  ingestionMethodId?: string;
  portalStatus?: string;
  sourceSystemBuyerId: string;
  sourceSystemId?: string;
  updatedDateTime?: string;
  displayIdentifierSearchValues?: string;
}